Move util_compare > util/compare
