+++ libmamba/tests/CMakeLists.txt
@@ -16,6 +16,7 @@ set(LIBMAMBA_TEST_SRCS
     # Utility library
     src/util/test_flat_set.cpp
     src/util/test_graph.cpp
+    src/util/test_compare.cpp
     # Implementation of version and matching specs
     src/specs/test_version.cpp
 
@@ -39,7 +40,6 @@ set(LIBMAMBA_TEST_SRCS
     src/core/test_virtual_packages.cpp
     src/core/test_util.cpp
     src/core/test_util_cast.cpp
-    src/core/test_util_compare.cpp
     src/core/test_util_string.cpp
     src/core/test_env_lockfile.cpp
     src/core/test_execution.cpp
